Windows SMBFS: allow mounting vhd/x images
The Windows SMBFS connector returns VHD/x image paths exposed by
the Cinder SMBFS volume driver.

This is fine for Nova, which will just use those image paths when
attaching volumes to instances.

The issue is that during backups/migrations, Cinder expects a raw
disk path, opening it and attempting to read volume data.

This change adds an argument to the connector init method, allowing
a raw disk path to be explicitly requested. If this argument is set,
we'll mount the image and bring the disk offline, returning the
raw disk path. This will allow direct IO operations to be performed
against the volume.

Related-Blueprint: windows-smb-backup

Change-Id: I20f791482fb0912772fa62d2949fa5becaec5675